requirements.txt
openai>=1.12.0
gradio>=4.0.0
browser-use>=0.1.35
python-dotenv>=1.0.0
